+++ src/cqspoolterrain/__init__.py
@@ -6,6 +6,13 @@ from .ControlPlatform import ControlPlatform
 from .SpoolCladding import SpoolCladding
 from .SpoolCladdingGreebled import SpoolCladdingGreebled
 from .SpoolCladdingGreebledUnique import SpoolCladdingGreebledUnique
-from .powerLine import (power_line_straight, power_line_curve, make_power_face, make_curved_connectors, make_curved_magnets)
+from .powerLine import (
+    power_line_straight, 
+    power_line_curve, 
+    make_power_face, 
+    make_curved_connectors, 
+    make_curved_magnets,
+    power_line_end
+)
 
 from .PowerStation import PowerStation